Design the Config Within the key Provider
