+++ mpesa-mobile-payment-impl/src/main/java/com/kwawingu/payments/session/MpesaSession.java
@@ -2,7 +2,6 @@ package com.kwawingu.payments.session;
 
 import com.kwawingu.payments.ApiEndpoint;
 import com.kwawingu.payments.Environment;
-import com.kwawingu.payments.SessionKeyGenerator;
 import com.kwawingu.payments.Market;
 import com.kwawingu.payments.Service;
 import com.kwawingu.payments.exception.SessionKeyUnavailableException;
